Update index.md
At the end of pull request #56 @bkeepers explained the thinking behind this template code of conduct. 

In the interest of being constructive in my criticism here is my proposal for a CoC that fits his aims. If @bkeepers wishes to lock this comment thread too pending the unlock of pull request #56 - then be my guest. 

-------------------------

My goal with this was to not single out any particular group as being an "acceptable target", to me even the concept of someone being an acceptable target is enough to make a convention feel actively hostile. 

I believe this captures as much of the original intent as possible without explicitly referring to any group. 

(tagging in @MrStonedOne @ammeep )