Removed protection for cisphobic and racist individuals
There should be no difference in policy between 'marginalized' and 'privileged' individuals, because offering protection to one group and not another based on their gender, gender identity and expression, sexual orientation, disability, mental illness, neuro(a)typicality, physical appearance, body size, race, age, regional discrimination, political or religious affiliation is defined as harassment.

There is no real way to determine whether someone is 'privileged' or not, and regardless should not be offered less protection because of it.
Reverse racism exists as regular racism, and there should be no protection for racists just because they're being racist against a 'privileged' ethnic group. It's still racism and should not be tolerated.